Move helper function from endpoints to crud module
